Feat: SSR runtime

+++ ci_tests.sh
@@ -17,9 +17,11 @@ if [ "$VERSION" = "$LATEST_VERSION" ]; then
     docker run --rm --name open-runtimes-formatter -v $(pwd):/mnt/code:rw open-runtimes/test-runtime sh -c "cd /mnt/code && $FORMATTER_PREPARE && $FORMATTER_CHECK"
     cd ../../
 
-    cd "tests/resources/functions/$RUNTIME_FOLDER"
-    docker run --rm --name open-runtimes-formatter -v $(pwd):/mnt/code:rw open-runtimes/test-runtime sh -c "cd /mnt/code && $FORMATTER_PREPARE && $FORMATTER_CHECK"
-    cd ../../../../
+    if [ -d "tests/resources/functions/$RUNTIME_FOLDER" ]; then
+        cd "tests/resources/functions/$RUNTIME_FOLDER"
+        docker run --rm --name open-runtimes-formatter -v $(pwd):/mnt/code:rw open-runtimes/test-runtime sh -c "cd /mnt/code && $FORMATTER_PREPARE && $FORMATTER_CHECK"
+        cd ../../../../
+    fi
 else
     echo "Skipping formatter. Formatter runs only in: $RUNTIME-$LATEST_VERSION"
 fi
